# apps/drive/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Xenode Internal API",
    "description": "🔒 **INTERNAL USE ONLY** — This API reference is for internal development and integration purposes. All endpoints require authentication via session cookie or API key bearer token.\n\nXenode is a zero-knowledge encrypted cloud storage platform. The backend never sees raw file contents or unencrypted keys. All E2EE operations (encrypt, decrypt, key wrapping) are performed client-side.",
    "version": "0.1.0",
    "contact": {
      "name": "Xenode Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/share/{token}/stream": {
      "post": {
        "tags": [
          "Link Shares"
        ],
        "summary": "Stream shared file",
        "description": "Returns a streaming URL for preview/playback of shared content. Returns direct R2 GET URLs lasting at most 300 seconds, rounded down to the link expiry. Responses are private, no-store. Downloads atomically consume the allowance; previews do not consume it.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream URL returned"
          }
        }
      }
    },
...
    "/api/direct-shares/{id}/stream": {
      "get": {
        "tags": ["Direct Shares"],
        "summary": "Stream direct-shared file",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Stream URL" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookieAuth": {
        "type": "apiKey", "in": "cookie", "name": "__Host-Xenode_admin_session",
        "description": "Drive Admin JWT cookie, validated against the current Admin record and session version. Development uses Xenode_admin_session. This is separate from user authentication."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Generate keys via `POST /api/keys`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie managed by better-auth. 7-day expiry, auto-refreshes every 24h."
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron job secret token (CRON_SECRET env var)."
      }
    },
...
  }
}